00-agent-agentcore/agent_files/requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents
strands-agents-tools
requests
twelvelabs
yt-dlp
